Calculates the weights required and plots graphs for training
The program predicts the weights to get the equation of the line separating
the two classes of output. It plots the graph for error as it is trained.
It also plots graph for training of AND, OR, NAND, NOR boolean gates and
demonstrates that this perceptron is insufficient for XOR gate